agentrun-browser-demo/requirements.txt
# LangChain 核心库
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# AgentRun SDK
agentrun-sdk[playwright,server]>=0.0.21

# 浏览器自动化
playwright>=1.40.0

# 环境变量管理
python-dotenv>=1.0.0

# Pydantic 用于数据验证
pydantic>=2.0.0
# LangChain + AgentRun SDK 浏览器自动化示例

# 核心依赖
agentrun-sdk[playwright,langchain]>=0.0.21
langchain>=0.3.0
langchain-community>=0.3.0

# HTTP 客户端
httpx>=0.27.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0

# 可选：如果你需要使用特定的模型
# openai>=1.0.0
# langchain-openai>=0.2.0
